Added navigation with drawer for navigation between screens. Added basic texts for home and login screen to show on screen.
