+++ App.tsx
@@ -1,38 +1,21 @@
-import { StatusBar } from "expo-status-bar";
+//React imports
 import React from "react";
-import { StyleSheet, Text, View, Button } from "react-native";
-import { Header } from "./components/Header.view/Header.component";
-import { storeData } from "./utils/config";
+import { Text, View } from "react-native";
+import { NavigationContainer } from "@react-navigation/native";
+import { createDrawerNavigator } from "@react-navigation/drawer";
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "pink",
-    alignItems: "center",
-    justifyContent: "center",
-    borderColor: "red",
-    borderWidth: 5,
-  },
-  bottomContainer: {
-    flex: 0.95,
-  },
-});
+//Screen imports
+import { Home } from "./components/Home.screen/Home.component";
+
+const Drawer = createDrawerNavigator();
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <Header />
-      <View style={styles.bottomContainer}>
-        <Text>NBC Meetup</Text>
-        <Button
-          onPress={() => {
-            storeData();
-          }}
-          title="Click Me"
-          color="#841584"
-        />
-        <StatusBar style="auto" />
-      </View>
-    </View>
+    <NavigationContainer>
+      <Drawer.Navigator initialRouteName="Home">
+        <Drawer.Screen name="Home" component={Home} />
+        {/* <Drawer.Screen name="" component={} /> */}
+      </Drawer.Navigator>
+    </NavigationContainer>
   );
 }
