+++ components/Header.view/Header.component.tsx
@@ -1,16 +1,21 @@
-import { Button, View, Image } from "react-native";
+import { Button, View, Image, Text } from "react-native";
 import React from "react";
 import { styles } from "./Header.style";
 import avatar from "../../assets/avatar.png";
+import { createDrawerNavigator } from "@react-navigation/drawer";
 
-export const Header = () => {
+export const Header = ({ navigation }) => {
   return (
     <View style={styles.container}>
       <Image source={avatar} style={styles.avatar}></Image>
+
+      {/* <Nav /> */}
       <Button
-        title="NAV"
+        title={"Nav"}
         onPress={() => {
           //Functionality to open up nav side bar
+          console.log("open navigation");
+          //navigation.toggleDrawer();
         }}
       ></Button>
     </View>
