+++ components/Header.view/Header.style.tsx
@@ -5,12 +5,9 @@ const windowHeight = Dimensions.get("window").height;
 export const styles = StyleSheet.create({
   container: {
     flexDirection: "row",
-    backgroundColor: "blue",
     alignItems: "center",
     justifyContent: "space-between",
     height: windowHeight / 15,
-    borderColor: "red",
-    borderWidth: 5,
     width: windowWidth,
     paddingHorizontal: windowWidth / 50,
   },
