+++ components/Home.screen/Home.component.tsx
@@ -1,0 +1,6 @@
+import { Text } from "react-native";
+import React from "react";
+
+export const Home = () => {
+  return <Text>HOME SCREEN</Text>;
+};
