+++ components/Login.screen/Login.component.tsx
@@ -1,0 +1,5 @@
+import { Text } from "react-native";
+
+export const Login = () => {
+  return <Text>LOG IN</Text>;
+};
