+++ components/Nav.view/Nav.component.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { createDrawerNavigator } from "@react-navigation/drawer";
+import { NavigationContainer } from "@react-navigation/native";
+import { Home } from "../Home.screen/Home.component";
+import { Login } from "../Login.screen/Login.component";
+
+NavigationContainer;
+
+const Drawer = createDrawerNavigator();
+export function Nav() {
+  return (
+    <Drawer.Navigator initialRouteName="Home">
+      <Drawer.Screen name="Home" component={Home} />
+      <Drawer.Screen name="Login" component={Login} />
+    </Drawer.Navigator>
+  );
+}
